Package names all lower cased.
