+++ src/main/java/entities/enemyfactory/ThirdBoss.java
@@ -1,4 +1,4 @@
-package entities.enemyFactory;
+package entities.enemyfactory;
 
 import entities.player.Character;
 import entities.player.Player;
