+++ src/main/java/usecases/CombatUseCase.java
@@ -1,4 +1,4 @@
-package useCases;
+package usecases;
 
 import java.util.Random;
