+++ src/main/java/usecases/DeleteGame.java
@@ -1,4 +1,4 @@
-package useCases;
+package usecases;
 
 import entities.GameStorage;
