+++ src/main/java/usecases/ExitGame.java
@@ -1,4 +1,4 @@
-package useCases;
+package usecases;
 
 import entities.TempEldenDisk;
