+++ src/main/java/usecases/SaveGame.java
@@ -1,4 +1,4 @@
-package useCases;
+package usecases;
 import entities.GameStorage;
 import java.io.IOException;
