+++ src/main/java/usecases/SaveGameGateway.java
@@ -1,4 +1,4 @@
-package useCases;
+package usecases;
 import entities.TempEldenDisk;
 
 public interface SaveGameGateway {
